Fix potential leak
If the newVal is never set on a child's content entry, we will just end
up leaking that value at the end of the function. This should ensure that
if the value is used, even once, we do not free it at the end of the
function.

However, if it is not used at least once, we will free it at the end
of the function call.